+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
-Nothing yet.
+### Added
+- Dark color theme
+- Persistence -- to save your preference if you have changed any setting
 
 
 ## [1.0.1] -- 2025-04-10